WIP refator transfer form to own component
